pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mengram-ai"
version = "2.6.2"
description = "Human-like memory for AI — semantic, episodic & procedural. Cognitive Profile, unified search, memory agents. Free open-source Mem0 alternative."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Ali Baizhanov"},
]
keywords = [
    "memory", "mengram", "knowledge-graph", "llm", "ai", "mcp",
    "second-brain", "rag", "semantic-search", "embeddings",
    "agents", "webhooks", "mem0-alternative", "langchain",
    "episodic-memory", "procedural-memory", "cognitive-profile",
    "contextual-memory", "langchain-memory",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
embeddings = ["sentence-transformers>=2.2"]
mcp = ["mcp>=1.0"]
api = ["fastapi>=0.100", "uvicorn>=0.20", "resend>=2.0"]
langchain = ["langchain-core>=0.2"]
crewai = ["crewai>=0.80"]
all = [
    "anthropic>=0.40",
    "openai>=1.0",
    "sentence-transformers>=2.2",
    "mcp>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "langchain-core>=0.2",
    "crewai>=0.80",
]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.urls]
Homepage = "https://github.com/alibaizhanov/mengram"
Repository = "https://github.com/alibaizhanov/mengram"
Issues = "https://github.com/alibaizhanov/mengram/issues"
Documentation = "https://mengram.io"

[project.scripts]
mengram = "cli:main"

[tool.setuptools.packages.find]
include = ["engine*", "api*", "cloud*", "integrations*"]

[tool.setuptools]
py-modules = ["mengram", "mengram_middleware", "cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
